+++ packages/base/src/types/ItemNavigationBehavior.ts
@@ -1,7 +1,7 @@
 /**
  * Different behavior for ItemNavigation.
  *
- * @class
+ * @readonly
  * @enum {string}
  * @public
  * @author SAP SE