+++ packages/fiori/src/UploadCollectionItem.ts
@@ -1,5 +1,4 @@
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
-import languageAware from "@ui5/webcomponents-base/dist/decorators/languageAware.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
@@ -57,8 +56,20 @@ import UploadCollectionItemCss from "./generated/themes/UploadCollectionItem.css
  * @implements sap.ui.webc.fiori.IUploadCollectionItem
  * @since 1.0.0-rc.7
  */
-@customElement("ui5-upload-collection-item")
-@languageAware
+@customElement({
+	tag: "ui5-upload-collection-item",
+	languageAware: true,
+	styles: [ListItem.styles, UploadCollectionItemCss],
+	template: UploadCollectionItemTemplate,
+	dependencies: [
+		...ListItem.dependencies,
+		Button,
+		Input,
+		Link,
+		Label,
+		ProgressIndicator,
+	],
+})
 
 /**
  * Fired when the file name is clicked.
@@ -246,25 +257,6 @@ class UploadCollectionItem extends ListItem {
 	 */
 
 	static i18nFioriBundle: I18nBundle;
-
-	static get styles() {
-		return [ListItem.styles, UploadCollectionItemCss];
-	}
-
-	static get template() {
-		return UploadCollectionItemTemplate;
-	}
-
-	static get dependencies() {
-		return [
-			...ListItem.dependencies,
-			Button,
-			Input,
-			Link,
-			Label,
-			ProgressIndicator,
-		];
-	}
 
 	static async onDefine() {
 		[UploadCollectionItem.i18nFioriBundle] = await Promise.all([