+++ packages/fiori/src/Wizard.ts
@@ -3,8 +3,6 @@ import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
-import languageAware from "@ui5/webcomponents-base/dist/decorators/languageAware.js";
-import fastNavigation from "@ui5/webcomponents-base/dist/decorators/fastNavigation.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import { getI18nBundle } from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
@@ -199,17 +197,30 @@ type StepInfo = {
  * @appenddocs sap.ui.webc.fiori.WizardStep
  * @public
  */
-@customElement("ui5-wizard")
-@languageAware
-@fastNavigation
+@customElement({
+	tag: "ui5-wizard",
+	languageAware: true,
+	fastNavigation: true,
+	renderer: litRender,
+	styles: WizardCss,
+	staticAreaStyles: WizardPopoverCss,
+	template: WizardTemplate,
+	staticAreaTemplate: WizardPopoverTemplate,
+	dependencies: [
+		WizardTab,
+		WizardStep,
+		ResponsivePopover,
+		Button,
+	],
+})
 
 /**
  * Fired when the step is changed by user interaction - either with scrolling,
  * or by clicking on the steps within the component header.
  *
  * @event sap.ui.webc.fiori.Wizard#step-change
- * @param {WizardStep} step The new step.
- * @param {WizardStep} previousStep The previous step.
+ * @param {sap.ui.webc.fiori.IWizardStep} step The new step.
+ * @param {sap.ui.webc.fiori.IWizardStep} previousStep The previous step.
  * @param {boolean} changeWithClick The step change occurs due to user's click or 'Enter'/'Space' key press on step within the navigation.
  * @public
  */
@@ -331,10 +342,6 @@ class Wizard extends UI5Element {
 		this._onStepResize = this.onStepResize.bind(this);
 	}
 
-	static get render() {
-		return litRender;
-	}
-
 	get classes() {
 		return {
 			popover: {
@@ -345,37 +352,12 @@ class Wizard extends UI5Element {
 		};
 	}
 
-	static get styles() {
-		return WizardCss;
-	}
-
-	static get staticAreaStyles() {
-		return WizardPopoverCss;
-	}
-
-	static get template() {
-		return WizardTemplate;
-	}
-
-	static get dependencies() {
-		return [
-			WizardTab,
-			WizardStep,
-			ResponsivePopover,
-			Button,
-		];
-	}
-
 	static async onDefine() {
 		Wizard.i18nBundle = await getI18nBundle("@ui5/webcomponents-fiori");
 	}
 
 	static get SCROLL_DEBOUNCE_RATE() {
 		return 25;
-	}
-
-	static get staticAreaTemplate() {
-		return WizardPopoverTemplate;
 	}
 
 	onExitDOM() {