+++ packages/fiori/src/types/MediaGalleryMenuHorizontalAlign.ts
@@ -1,5 +1,7 @@
 /**
- * @class
+ * Defines the horizontal alignment of the thumbnails menu of the <code>ui5-media-gallery</code> component.
+ *
+ * @readonly
  * @enum {string}
  * @public
  * @author SAP SE