+++ packages/fiori/src/types/UploadState.ts
@@ -1,7 +1,7 @@
 /**
  * Different types of UploadState.
  *
- * @class
+ * @readonly
  * @enum {string}
  * @public
  * @author SAP SE