+++ packages/main/src/types/TabContainerTabsPlacement.ts
@@ -1,9 +1,9 @@
 /**
  * Tabs placement of TabContainer.
  *
- * @class
+ * @readonly
  * @enum {string}
- * @public
+ * @private
  * @author SAP SE
  * @alias sap.ui.webc.main.types.TabContainerTabsPlacement
  */