+++ packages/playground/_stories/fiori/SideNavigation/SideNavigation.stories.ts
@@ -1,5 +1,7 @@
 import { html } from "lit-html";
-import type { Meta, StoryFn } from "@storybook/web-components";
+import { unsafeHTML } from "lit-html/directives/unsafe-html.js";
+import { ifDefined } from "lit-html/directives/if-defined.js";
+import type { Meta } from "@storybook/web-components";
 
 import argTypes, { componentInfo } from "./argTypes.js";
 import type { StoryArgsSlots } from "./argTypes.js";
@@ -7,67 +9,82 @@ import type { UI5StoryArgs } from "../../../types.js";
 
 import { DocsPage } from "../../../.storybook/docs";
 
-// @ts-ignore
 import type SideNavigation from "@ui5/webcomponents-fiori/dist/SideNavigation.js";
 
 const component = "ui5-side-navigation";
 
 export default {
-    title: "Fiori/SideNavigation",
-    component,
-    subcomponents: {'SideNavigationItem' : 'ui5-side-navigation-item', 'SideNavigationSubItem' : 'ui5-side-navigation-sub-item'},
-    parameters: {
-        docs: {
-          page: DocsPage({ ...componentInfo, component })
-        },
-    },
-    argTypes,
+	title: "Fiori/SideNavigation",
+	component,
+	subcomponents: {
+		SideNavigationItem: "ui5-side-navigation-item",
+		SideNavigationSubItem: "ui5-side-navigation-sub-item"
+	},
+	parameters: {
+		docs: {
+			page: DocsPage({ ...componentInfo, component })
+		},
+	},
+	argTypes,
 } as Meta<SideNavigation>;
 
-const Template: UI5StoryArgs<SideNavigation, StoryArgsSlots> = (args) => html`<div></div>`;
+const Template: UI5StoryArgs<SideNavigation, StoryArgsSlots> = (args) => {
+	return html`
+<ui5-side-navigation ?collapsed="${ifDefined(args.collapsed)}">
+	${args.header ? unsafeHTML(args.header) : null}
+	${unsafeHTML(args.default)}
+	${unsafeHTML(args.fixedItems)}
+</ui5-side-navigation>`;
+};
 
+export const Basic = Template.bind({});
+Basic.args = {
+	default: `<ui5-side-navigation-item text="Home" icon="home"></ui5-side-navigation-item>
+	<ui5-side-navigation-item text="People" expanded="" icon="group">
+		<ui5-side-navigation-sub-item text="From My Team"></ui5-side-navigation-sub-item>
+		<ui5-side-navigation-sub-item text="From Other Teams"></ui5-side-navigation-sub-item>
+	</ui5-side-navigation-item>
+	<ui5-side-navigation-item text="Locations" icon="locate-me" selected=""></ui5-side-navigation-item>
+	<ui5-side-navigation-item text="Events" icon="calendar">
+		<ui5-side-navigation-sub-item text="Local"></ui5-side-navigation-sub-item>
+		<ui5-side-navigation-sub-item text="Others"></ui5-side-navigation-sub-item>
+	</ui5-side-navigation-item>`,
+	fixedItems: `<ui5-side-navigation-item slot="fixedItems" text="Useful Links" icon="chain-link"></ui5-side-navigation-item>
+	<ui5-side-navigation-item slot="fixedItems" text="History" icon="history"></ui5-side-navigation-item>`
+};
+Basic.decorators = [
+	(story) => {
+		return html`<style>
+	ui5-side-navigation {
+		height: 600px;
+	}
 
-export const Template0: StoryFn = () => html`
-<style>
-		.ui5-side-nav {
-			height: 93%;
-		}
-		ui5-shellbar::part(root) {
-			padding-left: .5rem;
-		}
-	</style>
-	<h3>Side Navigation in Application</h3>
-	<div class="snippet" style="height: 40rem;">
-		<ui5-shellbar primary-title="UI5 Web Components" secondary-title="The Best Run SAP" show-co-pilot="">
-			<ui5-button icon="menu" slot="startButton" id="startButton"></ui5-button>
-		</ui5-shellbar>
-		<ui5-side-navigation class="ui5-side-nav">
-			<ui5-side-navigation-item text="Home" icon="home"></ui5-side-navigation-item>
-			<ui5-side-navigation-item text="People" expanded="" icon="group">
-				<ui5-side-navigation-sub-item text="From My Team"></ui5-side-navigation-sub-item>
-				<ui5-side-navigation-sub-item text="From Other Teams"></ui5-side-navigation-sub-item>
-			</ui5-side-navigation-item>
-			<ui5-side-navigation-item text="Locations" icon="locate-me" selected=""></ui5-side-navigation-item>
-			<ui5-side-navigation-item text="Events" icon="calendar">
-				<ui5-side-navigation-sub-item text="Local"></ui5-side-navigation-sub-item>
-				<ui5-side-navigation-sub-item text="Others"></ui5-side-navigation-sub-item>
-			</ui5-side-navigation-item>
-			<ui5-side-navigation-item slot="fixedItems" text="Useful Links" icon="chain-link"></ui5-side-navigation-item>
-			<ui5-side-navigation-item slot="fixedItems" text="History" icon="history"></ui5-side-navigation-item>
-		</ui5-side-navigation>
-		<script>
-			var sideNavigation = document.querySelector("ui5-side-navigation");
-			document.querySelector("#startButton").addEventListener("click", function(event) {
-				sideNavigation.collapsed = !sideNavigation.collapsed;
-			});
-		</script>
-	</div>
-`;
-Template0.parameters = {
+	ui5-shellbar::part(root) {
+		padding-inline-start: 0.5rem;
+	}
+</style>
+
+<ui5-shellbar primary-title="UI5 Web Components" secondary-title="The Best Run SAP" show-co-pilot>
+	<ui5-button icon="menu" slot="startButton" id="toggle"></ui5-button>
+</ui5-shellbar>
+
+${story()}
+
+<script>
+	const sidenav = document.querySelector("ui5-side-navigation");
+	document.getElementById("toggle").addEventListener("click", () => {
+		sidenav.toggleAttribute("collapsed");
+	});
+</script>`;
+	}
+]
+
+Basic.parameters = {
 	docs: {
 		story: {
-			// Opt-out of inline rendering
+			iframeHeight: "700px",
+			// Opt-out of inline rendering in Docs page
 			inline: false,
 		},
 	}
-};
+}