+++ packages/playground/_stories/main/Badge/Badge.stories.ts
@@ -1,75 +1,76 @@
 import { html } from "lit-html";
-import type { Meta, StoryFn } from "@storybook/web-components";
-
+import { ifDefined } from "lit-html/directives/if-defined.js";
+import { unsafeHTML } from "lit-html/directives/unsafe-html.js";
+import type { Meta } from "@storybook/web-components";
 import argTypes, { componentInfo } from "./argTypes.js";
 import type { StoryArgsSlots } from "./argTypes.js";
 import type { UI5StoryArgs } from "../../../types.js";
-
 import { DocsPage } from "../../../.storybook/docs";
-
-// @ts-ignore
 import type Badge from "@ui5/webcomponents/dist/Badge.js";
 
 const component = "ui5-badge";
 
 export default {
-    title: "Main/Badge",
-    component,
-    parameters: {
-        docs: {
-          page: DocsPage({ ...componentInfo, component })
-        },
-    },
-    argTypes,
+	title: "Main/Badge",
+	component,
+	parameters: {
+		docs: {
+			page: DocsPage({ ...componentInfo, component })
+		},
+	},
+	argTypes,
 } as Meta<Badge>;
 
-const Template: UI5StoryArgs<Badge, StoryArgsSlots> = (args) => html`<div></div>`;
+const Template: UI5StoryArgs<Badge, StoryArgsSlots> = (args) => {
+	return html`
+<ui5-badge
+	color-scheme="${ifDefined(args.colorScheme)}"
+	style="${ifDefined(args.style)}"
+>
+	${unsafeHTML(args.icon)}
+	${unsafeHTML(args.default)}
+</ui5-badge>`;
+};
 
+export const Basic = Template.bind({});
+Basic.args = {
+	colorScheme: "6",
+	icon: `<ui5-icon name="pending" slot="icon"></ui5-icon>`,
+	default: "Pending"
+};
 
-export const Template0: StoryFn = () => html`
-<h3>Basic Badge</h3>
-	<div class="snippet">
-		<ui5-badge color-scheme="1">available 1</ui5-badge>
-		<ui5-badge color-scheme="2">required 2</ui5-badge>
-		<ui5-badge color-scheme="3">3K</ui5-badge>
-		<ui5-badge color-scheme="4">bug 4</ui5-badge>
-		<ui5-badge color-scheme="5">in process 5</ui5-badge>
-		<ui5-badge color-scheme="6">in warehouse 6</ui5-badge>
-		<ui5-badge color-scheme="7">7\$</ui5-badge>
-		<ui5-badge color-scheme="8">solution provided 8</ui5-badge>
-		<ui5-badge color-scheme="9">pending release 9</ui5-badge>
-		<ui5-badge color-scheme="10">customer action 10</ui5-badge>
-		<ui5-badge style="width:200px;">This would truncate as it is too long</ui5-badge>
-	</div>
-`;
+export const Truncating = Template.bind({});
+Truncating.args = {
+	default: "This would truncate as it is too long",
+	style: "width: 200px",
+};
 
+const getIconHTML = (name: string): string => `<ui5-icon name="${name}" slot="icon"></ui5-icon>`;
+const AllColorSchemesBadges = [
+	{ icon: getIconHTML("accept"), default: "" },
+	{ icon: getIconHTML("sap-ui5"), default: "" },
+	{ icon: getIconHTML("add-equipment"), default: "In progress" },
+	{ icon: getIconHTML("lab"), default: "" },
+	{ icon: getIconHTML("email-read"), default: "" },
+	{ icon: "", default: "Pending" },
+	{ icon: getIconHTML("lightbulb"), default: "New idea" },
+	{ icon: getIconHTML("locked"), default: "Locked" },
+	{ icon: getIconHTML("flight"), default: "En route" },
+	{ icon: "", default: "Archived" },
+];
 
-export const Template1: StoryFn = () => html`
-<h3>Badge with Icon</h3>
-	<div class="snippet">
-		<ui5-badge color-scheme="1">
-				<ui5-icon name="accept" slot="icon"></ui5-icon>done
-		</ui5-badge>
-		<ui5-badge color-scheme="2">
-				<ui5-icon name="sap-ui5" slot="icon"></ui5-icon>
-		</ui5-badge>
-		<ui5-badge color-scheme="3">
-				<ui5-icon name="add-equipment" slot="icon"></ui5-icon>in process
-		</ui5-badge>
-		<ui5-badge color-scheme="4">
-				<ui5-icon name="lab" slot="icon"></ui5-icon>
-		</ui5-badge>
-		<ui5-badge color-scheme="5">
-				<ui5-icon name="email-read" slot="icon"></ui5-icon>
-		</ui5-badge>
-		<ui5-badge color-scheme="6">
-				<ui5-icon name="pending" slot="icon"></ui5-icon>pending
-		</ui5-badge>
-		<ui5-badge color-scheme="7">
-				<ui5-icon name="lightbulb" slot="icon"></ui5-icon>
-		</ui5-badge>
-		<ui5-badge color-scheme="8">
-				<ui5-icon name="locked" slot="icon"></ui5-icon>
-		</ui5-badge>
-	</div>
-`;
+export const AllColorSchemes = Template.bind({});
+AllColorSchemes.decorators = [
+	(story, ctx) => {
+		return html`
+			${AllColorSchemesBadges.map((badge, i) => {
+			return story({
+				args: {
+					colorScheme: ctx.args.colorScheme || (i + 1).toString(),
+					icon: ctx.args.icon || badge.icon,
+					default: ctx.args.default || badge.default,
+				}
+			});
+		})}`;
+	}
+];