+++ packages/playground/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.12.0-rc.0](https://github.com/SAP/ui5-webcomponents/compare/v1.11.0...v1.12.0-rc.0) (2023-03-09)
+
+**Note:** Version bump only for package @ui5/webcomponents-playground
+
+
+
+
+
+# [1.11.0](https://github.com/SAP/ui5-webcomponents/compare/v1.11.0-rc.4...v1.11.0) (2023-03-06)
+
+
+### Bug Fixes
+
+* **playgrouond:** process docs path os agnostic ([#6618](https://github.com/SAP/ui5-webcomponents/issues/6618)) ([7bd2dc0](https://github.com/SAP/ui5-webcomponents/commit/7bd2dc0aaa9132c3418becc9a3fa8af9bef9ef4b))
+
+
+
+
+
+# [1.11.0-rc.4](https://github.com/SAP/ui5-webcomponents/compare/v1.11.0-rc.3...v1.11.0-rc.4) (2023-03-02)
+
+
+### Bug Fixes
+
+* **playground:** scope components inside stories ([#6552](https://github.com/SAP/ui5-webcomponents/issues/6552)) ([df9c75a](https://github.com/SAP/ui5-webcomponents/commit/df9c75a2038559e1ef50b3fba79aa4c00c619f1e))
+
+
+
+
+
 # [1.11.0-rc.3](https://github.com/SAP/ui5-webcomponents/compare/v1.11.0-rc.2...v1.11.0-rc.3) (2023-02-23)